+++ internal/app/balance_service_app/handlers/account_handler/request_models/refill_message.go
@@ -1,7 +1,7 @@
 package request_models
 
 type RefillMessage struct {
-	UserID  int64   `json:"user_id"`
+	UserID  int     `json:"user_id"`
 	Sum     float64 `json:"sum"`
 	Comment string  `json:"comment"`
 }
