+++ internal/app/balance_service_app/handlers/report_handler/request_models/fin_report_message.go
@@ -1,6 +1,6 @@
 package request_models
 
 type FinReportMessage struct {
-	Year  int64 `json:"year"`
-	Month int64 `json:"month"`
+	Year  int `json:"year"`
+	Month int `json:"month"`
 }
