+++ internal/app/balance_service_app/report/report.go
@@ -1,6 +1,6 @@
 package report
 
 type FinanceReport struct {
-	ServiceType int64
+	ServiceType int
 	Sum         float64
 }
